Add files via upload
The goal of this project is to analyse the collected data and build a model that predicts whether an employee will leave the company. By predicting employees likely to quit, we can identify contributing factors and take proactive measures to increase employee retention, which is beneficial considering the time and expense involved in hiring new employees.